Move classes to their own package
